style: set programlisting, note, important, informaltable to css style, add docbook chapters
this will allow users to easily determine the code blocks around
the shell code in the quickstart guides, detect hints and important
notes better, as well identify other 'screen' formatted blocks.

the informaltable requires a rework within docbook xml itsself -
currently, it's used for the colored example, as well as the
directive description below. both should get their seperated
div style somewhere in the future.

the docbook generated toc, as well as the sections now contain
section/chapter numbers, which allows easier identification of
the sections when reading through, as well as keep the toc in
mind.